Refactoring, get them into the popup
